Split PageCreateModal into button and modal
