+++ views/layout/layout.html
@@ -143,7 +143,7 @@
       {% endif %}
 
       <li class="nav-item">
-        <div id="page-create-modal"></div>
+        <div id="header-page-create-modal"></div>
       </li>
       <li class="nav-item nav-divider">
       </li>
